Implement sub-agent architecture for data quality management
- Added `analyst.py` to handle analyst agent operations, including JSON extraction and report saving.
- Introduced `auditor.py` for auditing the knowledge base, generating reports, and saving findings.
- Created `curator.py` to manage search ranking and URL ingestion, including report initialization and updates.
- Developed `fixer.py` for correcting data quality issues based on auditor reports, including human approval for plans.
- Implemented `researcher.py` to conduct research on identified gaps, updating reports with search results.
- Established `tools.py` with utility functions for JSON extraction, report loading, human approval, and report management.
