+++ sub_agents/curator.py
@@ -7,6 +7,7 @@ import json
 import os
 import re
 from state import AgentState
+from tools import initialize_curator, update_curator_report
 
 def _extract_and_clean_json_curator(llm_output: str) -> dict:
     """
@@ -50,221 +51,6 @@ def _extract_and_clean_json_curator(llm_output: str) -> dict:
         logger.error(status)
         raise ValueError(status)
     
-@tool
-def initialize_curator(timestamp: str) -> dict:
-    """
-    Initializes the curator's report.
-    This tool should be called once at the beginning of the curator's workflow.
-    It reads the researcher's report, creates the initial structure for the curator_report.json,
-    and returns the report_id and the list of searches to be populated into the AgentState.
-    """
-    import logging
-    logger = logging.getLogger('KnowledgeAgent')
-
-    status = f"Initialize curator tool called, attempting to load researcher_report.json ---"
-    print(status)
-    logger.info(status)
-
-    try:
-        researcher_report_str = load_report('researcher_report.json')
-    except Exception as e:
-        status = f"[ERROR] Error loading researcher report: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    if "No report found" in researcher_report_str:
-        status = f"[ERROR] Researcher report not found."
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    status = f"Loaded researcher report:\n{researcher_report_str}"
-    print(status)
-    logger.info(status)
-
-    researcher_report = json.loads(researcher_report_str)
-
-    report_id = f"cur_{timestamp.replace('-', '').replace(':', '').replace('T', '_').split('.')[0]}"
-    status = f"Generated curator report ID: {report_id}"
-    print(status)
-    logger.info(status)
-
-    searches_todo = []
-    for gap in researcher_report.get("gaps", []):
-        for search in gap.get("searches", []):
-            searches_todo.append({
-                "search_id": search["search_id"],
-                "rationale": search["rationale"],
-                "parameters": search["parameters"],
-                "results": search["results"]
-            })
-    status = f"Compiled {len(searches_todo)} searches to be curated."
-    print(f"[INFO] {status}")
-    logger.info(status)
-
-    new_report = {
-        "report_id": report_id,
-        "urls_for_ingestion": [],
-        "url_ingestion_status": [],
-    }
-
-    status = f"Initialized new curator report:\n{json.dumps(new_report, indent=2)}"
-    print(status)
-    logger.info(status)
-
-    filepath = "state/curator_report.json"
-
-    status = f"Loading curator reports into memory from {filepath}"
-    print(status)
-    logger.info(status)
-    try:
-        with open(filepath, "r") as f:
-            file_data = json.load(f)
-    except Exception as e:
-        status = f"[ERROR] Error reading {filepath}: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    file_data = {"reports": []}
-    status = f"Loading curator reports into data structure"
-    print(status)
-    logger.info(status)
-    try:
-        if os.path.exists(filepath) and os.path.getsize(filepath) > 0:
-            with open(filepath, "r") as f:
-                file_data = json.load(f)
-    except Exception as e:
-        logger.error(f"Error loading {filepath} into data structure: {e}")
-        print(f"Error loading {filepath} into data structure: {e}")
-        raise ToolException(f"Error loading {filepath} into data structure: {e}")
-
-    status = f"Loaded curator reports from {filepath} into data structure"
-    print(status)
-    logger.info(status)
-
-    status = f"Appending new report to curator reports data structure"
-    print(status)
-    logger.info(status)
-
-    try:
-        file_data["reports"].append(new_report)
-        status = f"Successfully appended new report to curator reports data structure"
-        print(status)
-        logger.info(status)
-    except Exception as e:
-        status = f"[ERROR] Error appending new report to curator reports data structure: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    status = f"Writing updated curator reports to {filepath}"
-    print(status)
-    logger.info(status)
-    try:
-        with open(filepath, "w") as f:
-            json.dump(file_data, f, indent=2)
-        status = f"Successfully wrote updated curator reports to {filepath}"
-        print(status)
-        logger.info(status)
-    except Exception as e:
-        status = f"[ERROR] Error writing updated curator reports to {filepath}: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    return {
-        "curator_report_id": report_id,
-        "curator_searches_todo": searches_todo
-    }
-
-@tool
-def update_curator_report(report_id: str, current_search: dict, job: str, results: list) -> str:
-    """
-    Updates the curator's report with the job results (urls_for_ingestion, or url_ingestion_status) for a single search.
-    """
-    import logging
-    logger = logging.getLogger('KnowledgeAgent')
-
-    status = f"Updating curator report {report_id} with new results for: {job}.\n{results}"
-    print(status)
-    logger.info(status)
-
-    filepath = "state/curator_report.json"
-    try:
-        with open(filepath, "r") as f:
-            file_data = json.load(f)
-    except Exception as e:
-        status = f"[ERROR] Could not read or parse {filepath}. Error: {str(e)}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-    # Find the report index
-    report_list = file_data.get("reports", [])
-    report_index = next((i for i, r in enumerate(report_list) if r.get("report_id") == report_id), None)
-    if report_index is None:
-        status = f"[ERROR] Report with ID {report_id} not found."
-        print(status)
-        logger.error(status)
-        raise ToolException(status)
-
-    # Update the job results
-
-    status = f"Updating {report_id} in the data structure with the new results."
-    print(status)
-    logger.info(status)
-    try:
-        file_data['reports'][report_index][job].append(results)
-    except Exception as e:
-        status = f"[ERROR] Error updating {report_id} in the data structure: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-    status = f"Inserted {len(results)} results to {report_id} in the data structure."
-    print(status)
-    logger.info(status)
-
-    # Write back to file
-    try:
-        with open(filepath, "w") as f:
-            json.dump(file_data, f, indent=2)
-        status = f"Successfully wrote updated data for report {report_id} to {filepath}"
-        logger.info(status)
-        print(status)
-        return status
-    except Exception as e:
-        status = f"[ERROR] Failed to write to {filepath}: {e}"
-        logger.error(status)
-        print(status)
-        raise ToolException(status)
-
-@tool
-def load_report(filename: str) -> str:
-    """Loads the most recent report from a file in the state directory."""
-    import logging
-    logger = logging.getLogger('KnowledgeAgent')
-
-    filepath = f"state/{filename}"
-    if not os.path.exists(filepath) or os.path.getsize(filepath) == 0:
-        status = f"[ERROR] No report found at {filepath}."
-        print(status)
-        logger.warning(status)
-        return status
-    with open(filepath, "r") as f:
-        data = json.load(f)
-    if "reports" in data and isinstance(data["reports"], list) and data["reports"]:
-        status = f"Loaded report from {filepath}: {json.dumps(data['reports'][-1], indent=2)}"
-        print(status)
-        logger.info(status)
-        return json.dumps(data["reports"][-1])
-    else:
-        status = "[ERROR] No reports found in the file."
-        print(status)
-        logger.error(status)
-        return status
-
 async def curator_agent_node(state: AgentState):
     logger = state['logger']
