+++ sub_agents/fixer.py
@@ -7,55 +7,8 @@ import json
 import os
 import re
 from state import AgentState
+from tools import load_report, human_approval, _extract_and_clean_json
 
-@tool
-def load_report(filename: str) -> str:
-    """Loads the most recent report from a file in the state directory."""
-    import logging
-    logger = logging.getLogger('KnowledgeAgent')
-
-    filepath = f"state/{filename}"
-    if not os.path.exists(filepath) or os.path.getsize(filepath) == 0:
-        status = f"[ERROR] No report found at {filepath}."
-        print(status)
-        logger.warning(status)
-        return status
-    with open(filepath, "r") as f:
-        data = json.load(f)
-    if "reports" in data and isinstance(data["reports"], list) and data["reports"]:
-        status = f"Loaded report from {filepath}: {json.dumps(data['reports'][-1], indent=2)}"
-        print(status)
-        logger.info(status)
-        return json.dumps(data["reports"][-1])
-    else:
-        status = "[ERROR] No reports found in the file."
-        print(status)
-        logger.error(status)
-        return status
-
-@tool
-def human_approval(plan: str) -> str:
-    """
-    Asks for human approval for a given plan.
-    The plan is a string that describes the actions to be taken.
-    Returns 'approved' or 'denied'.
-    """
-    import logging
-    logger = logging.getLogger('KnowledgeAgent')
-
-    status = f"PROPOSED PLAN:\n{plan}"
-    print(f"\n{status}")
-    logger.info(f"{status}")
-    response = input("Do you approve this plan? (y/n): ").lower()
-    if response == 'y':
-        status = "Approved."
-        print(status)
-        logger.info(status)
-        return "approved"
-    status = "Denied."
-    print(status)
-    logger.info(status)
-    return "denied"
 
 async def fixer_agent_node(state: AgentState):
     print("--- Running Fixer Agent ---")
@@ -92,7 +45,7 @@ def save_fixer_report_node(state: AgentState):
     logger.info(status)
 
     try:
-        report_json = _extract_and_clean_json_fixer(final_message_from_agent.content)
+        report_json = _extract_and_clean_json(final_message_from_agent.content)
         if 'report_id' not in report_json:
             report_json['report_id'] = state.get('fixer_report_id', 'unknown_id')
         save_fixer_report.invoke({"fixer_report": json.dumps(report_json)})
